requirements.txt
langchain
streamlit
python-dotenv
assemblyai
chromadb
sentence-transformers
langchain-community
yt-dlp
langchain-core 
langchain-upstage